requirements.txt
ccxt
pandas
ta
python-dotenv
matplotlib
